moved IT8951 to within waveshare display class renamed to be waveshare_epd.it8951
