+++ src/omni_epd/displays/waveshare_display.py
@@ -380,3 +380,60 @@ class WaveshareMultiColorDisplay(WaveshareDisplay):
             image = self._filterImage(image)
 
         self._device.display(self._device.getbuffer(image))
+
+class IT8951Display(VirtualEPD):
+    """
+    This will communicate with IT8951 type displays utilzing the driver built by GregDMeyer
+    https://github.com/GregDMeyer/IT8951
+    """
+
+    pkg_name = WAVESHARE_PKG
+    it8951_pkg_name = 'IT8951'
+    it8951_constants = None
+
+    def __init__(self, deviceName, config):
+        super().__init__(deviceName, config)
+
+        self.logger = logging.getLogger(__name__)
+
+        # load the IT8951.display package and create the object
+        deviceObj = self.load_display_driver(self.it8951_pkg_name, "display")
+        self._device = deviceObj.AutoEPDDisplay(vcom=self._getfloat_device_option('vcom', -2.06),
+                                                spi_hz=self._getint_device_option('spi_hz', 24000000),
+                                                rotate=self._get_device_option('rotate', None))
+
+        # set the width and height
+        self.width = self._device.width
+        self.height = self._device.height
+
+        # load the from IT8951.constants
+        self.it8951_constants = self.load_display_driver(self.it8951_pkg_name, "constants")
+
+    @staticmethod
+    def get_supported_devices():
+        # same type for all it8951 displays
+        return [f"{WAVESHARE_PKG}.it8951"]
+
+    def prepare(self):
+        self._device.epd.run()
+
+    def _display(self, image):
+        # keep this as it applies any filters defined in INI file
+        image = self._filterImage(image)
+
+        self.clear()  # not sure if this is needed, was part of example
+
+        dims = (self.width, self.height)
+        image.thumbnail(dims)
+
+        paste_coords = [dims[i] - image.size[i] for i in (0, 1)]  # align image with bottom of display
+
+        # write image to display
+        self._device.frame_buf.paste(image, paste_coords)
+        self._device.draw_full(self.it8951_constants.DisplayModes.GC16)
+
+    def sleep(self):
+        self._device.epd.sleep()
+
+    def clear(self):
+        self._device.clear()
